Revert "Proper physics"
